+++ packages/cli/commands/annotate-execution-points.ts
