+++ packages/cli/main.ts
@@ -2,7 +2,7 @@ import "./commands/fetch-comments";
 import "./commands/session";
 //import "./commands/sources";
 import "./commands/version";
-import "./commands/annotateExecutionPoints";
+import "./commands/annotate-execution-points";
 
 import { program } from "commander";
